+++ src/fabric/_types.ts
@@ -124,6 +124,7 @@ export type FabricApiItemType =
 	| "LivySessions"						//	Folder for Livy sessions.	
 	| "LivySession"							//	A Livy session.
 	| "SqlDatabaseMirroring"
+	| "AzureDatabricksStorage"
 	;
 
 export enum FabricApiWorkspaceType {