+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change Log
 
+**v2.23.1**:
+- added `Definition` to all item types
+- updated API Specs for auto-complete
+- added support for new item `AzureDatabricksStorage`
+
 **v2.23.0**:
 - fixed issue with Drag&Drop of workspace roles
 - update sidebar icon to make it more distinctive